#6001 feat(refactor): flatten component structure
